+++ section_code.php
@@ -44,16 +44,27 @@ if(isset($_POST['btn-update'])){
     }
 }
 
-if(isset($_POST['deleteSectionBtn'])){
+if (isset($_POST['deleteSectionBtn'])) {
     $delete_id = $_POST['delete_id'];
-    
-    $query = "DELETE FROM sections WHERE id='$delete_id' ";
-    $query_run = mysqli_query($conn,$query);
 
-    if($query_run){
-        $_SESSION['status'] = "Section deleted successfully";
-		header("location: student_section.php");
+    $check_usage_query = "SELECT COUNT(*) AS usage_count FROM assign_students WHERE section_id = '$delete_id'";
+    $check_usage_result = mysqli_query($conn, $check_usage_query);
+    if ($check_usage_result) {
+        $usage_count = mysqli_fetch_assoc($check_usage_result)['usage_count'];
+        if ($usage_count > 0) {
+            $_SESSION['notification'] = "Section deletion failed! This section is in use cannot be deleted.";
+            header("location: student_section.php");
+        } else {
+            $query = "DELETE FROM sections WHERE id='$delete_id' ";
+            $query_run = mysqli_query($conn, $query);
+            if ($query_run) {
+                $_SESSION['SuccessMessage'] = "Section deleted successfully.";
+                header("location: student_section.php");
+            } else {
+                $_SESSION['notification'] = "Something went wrong!";
+                header("location: student_section.php");
+            }
+        }
     }
-
 }
 ?>
